hardware wallets: show address on device also from tx dialog
