+++ electrum/plugins/digitalbitbox/qt.py
@@ -29,26 +29,23 @@ class Plugin(DigitalBitboxPlugin, QtPluginBase):
 
     @only_hook_if_libraries_available
     @hook
-    def receive_menu(self, menu, addrs, wallet: Abstract_Wallet):
-        if type(wallet) is not Standard_Wallet:
-            return
-
-        keystore = wallet.get_keystore()
-        if type(keystore) is not self.keystore_class:
-            return
-
+    def receive_menu(self, menu, addrs, wallet):
         if not self.is_mobile_paired():
             return
+        if len(addrs) != 1:
+            return
+        if wallet.get_txin_type(addrs[0]) != 'p2pkh':
+            return
+        self._add_menu_action(menu, addrs[0], wallet)
 
-        if len(addrs) == 1:
-            addr = addrs[0]
-            if wallet.get_txin_type(addr) != 'p2pkh':
-                return
-
-            def show_address():
-                keystore.thread.add(partial(self.show_address, wallet, addr, keystore))
-
-            menu.addAction(_("Show on {}").format(self.device), show_address)
+    @only_hook_if_libraries_available
+    @hook
+    def transaction_dialog_address_menu(self, menu, addr, wallet):
+        if not self.is_mobile_paired():
+            return
+        if wallet.get_txin_type(addr) != 'p2pkh':
+            return
+        self._add_menu_action(menu, addr, wallet)
 
     @hook
     def init_wallet_wizard(self, wizard: 'QENewWalletWizard'):
