+++ electrum/plugins/jade/qt.py
@@ -9,6 +9,7 @@ from electrum.wallet import Standard_Wallet
 
 from electrum.hw_wallet.qt import QtHandlerBase, QtPluginBase
 from electrum.hw_wallet import plugin
+from electrum.hw_wallet.plugin import only_hook_if_libraries_available
 from electrum.gui.qt.wizard.wallet import WCScriptAndDerivation, WCHWUnlock, WCHWXPub, WCHWUninitialized
 
 from .jade import JadePlugin
@@ -24,16 +25,21 @@ class Plugin(JadePlugin, QtPluginBase):
     def create_handler(self, window):
         return Jade_Handler(window)
 
-    @plugin.only_hook_if_libraries_available
+    @only_hook_if_libraries_available
     @hook
     def receive_menu(self, menu, addrs, wallet):
+        if len(addrs) != 1:
+            return
         if type(wallet) is not Standard_Wallet:
             return
-        keystore = wallet.get_keystore()
-        if type(keystore) == self.keystore_class and len(addrs) == 1:
-            def show_address():
-                keystore.thread.add(partial(self.show_address, wallet, addrs[0]))
-            menu.addAction(_("Show on Jade"), show_address)
+        self._add_menu_action(menu, addrs[0], wallet)
+
+    @only_hook_if_libraries_available
+    @hook
+    def transaction_dialog_address_menu(self, menu, addr, wallet):
+        if type(wallet) is not Standard_Wallet:
+            return
+        self._add_menu_action(menu, addr, wallet)
 
     @hook
     def init_wallet_wizard(self, wizard: 'QENewWalletWizard'):
